Tests/LibWeb/Text/expected/grid/zero-flex-row-overflow-hidden-grid-item.txt

root: top=0 height=35
outer: top=0 height=34
button: top=0 height=34
collapsible: top=34 height=0
clip: top=34 height=0
list: top=35 height=104
after: top=35 height=0
